Add a SamplesLikeTable type
